# libco/aarch64: Fix vector registers handling and optimizations