Originally submitted as byuu/higan#77.

---

The current implementation accidentally saved and restored registers x8-x15 instead of d8-d15. I'm not aware of this causing any issues in bsnes/higan, but hey, might as well fix it.

Additionally, I used [this simple benchmark](https://github.com/Sintendo/libco/blob/aarch64/-test/test_timing.cpp) to see if it's possible to get this to run slightly faster. Here's some benchmark results from the ARM64 CPUs I happen to have access to:

| CPU | Before (s) | After (s) |
|:-------------|-------------:|-----:|
| AWS Gravitron | 26.4 | 18.1 |
| BCM2837 (RPi 3B) | 44.4 | 36.4 |
| BCM2711 (RPi 4B) | 35.4 | 28.1 |

Impact on emulation speed is very limited; I saw maybe a 1-2 FPS increase in bsnes using the most accurate settings on my RPi4.