@Kawa-oneechan 2020-03-23
Looks good to me, says the one person here least likely to understand assembler code at this level. Anyone else?